# Conflicts:
#	src/ModelViewer.h
#	win32/vs2015/pioneer.vcxproj.filters

+++ data/libs/SpaceStation.lua
@@ -1,8 +1,9 @@
--- Copyright © 2008-2014 Pioneer Developers. See AUTHORS.txt for details
+-- Copyright © 2008-2016 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 local SpaceStation = import_core("SpaceStation")
 local Event = import("Event")
+local Rand = import("Rand")
 local Space = import("Space")
 local utils = import("utils")
 local ShipDef = import("ShipDef")
@@ -14,10 +15,22 @@ local Model = import("SceneGraph.Model")
 local ModelSkin = import("SceneGraph.ModelSkin")
 local Serializer = import("Serializer")
 local Equipment = import("Equipment")
+local Faction = import("Faction")
+local Lang = import("Lang")
+local l = Lang.GetResource("ui-core")
 
 --
 -- Class: SpaceStation
 --
+
+
+function SpaceStation:Constructor()
+	-- Use a variation of the space station seed itself to ensure consistency
+	local rand = Rand.New(util.hash_random(self.seed .. '-techLevel', 2^31)-1)
+	local techLevel = rand:Integer(1, 6) + rand:Integer(0,6)
+	self:setprop("techLevel", techLevel)
+end
+
 
 local equipmentStock = {}
 
@@ -204,41 +217,189 @@ local isPlayerShip = function (def) return def.tag == "SHIP" and def.basePrice >
 local groundShips = utils.build_array(utils.filter(function (k,def) return isPlayerShip(def) and def.equipSlotCapacity.atmo_shield > 0 end, pairs(ShipDef)))
 local spaceShips  = utils.build_array(utils.filter(function (k,def) return isPlayerShip(def) end, pairs(ShipDef)))
 
+
+-- Dynamics of ship adverts in ShipMarket --
+--------------------------------------------
+-- N(t) = Number of ads, lambda = decay constant:
+--    d/dt N(t) = prod - lambda * N
+-- and equilibrium:
+--    dN/dt = 0 = prod - lambda * N_equil
+-- and solution (if prod=0), with N_0 initial number:
+--    N(t) = N_0 * exp(-lambda * t)
+-- with tau = half life, i.e. N(tau) = 0.5*N_0 we get:
+--    0.5*N_0 = N_0*exp(-lambda * tau)
+-- else, with production:
+--   N(t) = prod/lambda - N_0*exp(-lambda * t)
+-- We want to have N_0 = N_equil, since ShipMarket should spawn in equilibrium
+
+-- Average number of ship for sale for station
+local function N_equilibrium(station)
+	local pop = station.path:GetSystemBody().parent.population -- E.g. Earth=7, Mars=0.3
+	local pop_bonus = 9 * math.log(pop*0.45 + 1)       -- something that gives resonable result
+	if station.type == "STARPORT_SURFACE" then
+		pop_bonus = pop_bonus * 1.5
+	end
+
+	return 2 + pop_bonus
+end
+
+-- add num random ships for sale to station ShipMarket
+local function addRandomShipAdvert(station, num)
+	for i=1,num do
+		local avail = station.type == "STARPORT_SURFACE" and groundShips or spaceShips
+		local def = avail[Engine.rand:Integer(1,#avail)]
+		local model = Engine.GetModel(def.modelName)
+		local pattern = model.numPatterns ~= 0 and Engine.rand:Integer(1,model.numPatterns) or nil
+		addShipOnSale(station, {
+			def     = def,
+			skin    = ModelSkin.New():SetRandomColors(Engine.rand):SetDecal(def.manufacturer),
+			pattern = pattern,
+			label   = Ship.MakeRandomLabel(),
+		})
+	end
+end
+
+local function createShipMarket (station)
+	shipsOnSale[station] = {}
+
+	local shipAdsToSpawn = Engine.rand:Poisson(N_equilibrium(station))
+	addRandomShipAdvert(station, shipAdsToSpawn)
+end
+
 local function updateShipsOnSale (station)
 	if not shipsOnSale[station] then shipsOnSale[station] = {} end
 	local shipsOnSale = shipsOnSale[station]
 
-	local toAdd, toRemove = 0, 0
-	if #shipsOnSale == 0 then
-		toAdd = Engine.rand:Integer(20)
-	elseif Engine.rand:Integer(1) > 0 then
-		toAdd = 1
-	elseif #shipsOnSale > 0 then
-		toRemove = 1
-	else
-		return
-	end
+	local tau = 7*24                              -- half life of a ship advert in hours
+	local lambda = 0.693147 / tau                 -- removal probability= ln(2) / tau
+	local prod = N_equilibrium(station) * lambda  -- creation probability
 
-	if toAdd > 0 then
-		local avail = station.type == "STARPORT_SURFACE" and groundShips or spaceShips
-		for i=1,toAdd do
-			local def = avail[Engine.rand:Integer(1,#avail)]
-			local model = Engine.GetModel(def.modelName)
-			local pattern = model.numPatterns ~= 0 and Engine.rand:Integer(1,model.numPatterns) or nil
-			addShipOnSale(station, {
-				def     = def,
-				skin    = ModelSkin.New():SetRandomColors(Engine.rand):SetDecal(def.manufacturer),
-				pattern = pattern,
-				label   = Ship.MakeRandomLabel(),
-			})
+	-- remove with decay rate lambda. Call ONCE/hour for each ship advert in station
+	for ref,ad in pairs(shipsOnSale) do
+		if Engine.rand:Number(0,1) < lambda then  -- remove one random ship (sold)
+			removeShipOnSale(station, Engine.rand:Integer(1,#shipsOnSale))
 		end
 	end
 
-	if toRemove > 0 then
-		removeShipOnSale(station, Engine.rand:Integer(1,#shipsOnSale))
+	-- spawn a new ship adverts, call for each station
+	if Engine.rand:Number(0,1) <= prod then
+		addRandomShipAdvert(station, 1)
 	end
 
+	if prod > 1 then print("Warning: ShipMarket not in equilibrium") end
+
 	Event.Queue("onShipMarketUpdate", station, shipsOnSale)
+end
+
+
+--
+-- Attribute: lawEnforcedRange
+--
+--   The distance, in meters, at which a station upholds the law,
+--   (is 100 km for all at the moment)
+--
+-- Availability:
+--
+--   2015 September
+--
+-- Status:
+--
+--   experimental
+--
+SpaceStation.lawEnforcedRange = 100000
+
+
+local police = {}
+
+--
+-- Method: LaunchPolice
+--
+-- Launch station police
+--
+-- > station:LaunchPolice(targetShip)
+--
+-- Parameters:
+--
+--   targetShip - the ship to intercept
+--
+-- Availability:
+--
+--   2015 September
+--
+-- Status:
+--
+--   experimental
+--
+function SpaceStation:LaunchPolice(targetShip)
+	if not targetShip then error("Ship targeted invalid") end
+
+	-- if no police created for this station yet:
+	if not police[self] then
+		police[self] = {}
+		-- decide how many to create
+		local lawlessness = Game.system.lawlessness
+		local maxPolice = math.min(9, self.numDocks)
+		local numberPolice = math.ceil(Engine.rand:Integer(1,maxPolice)*(1-lawlessness))
+		local shiptype = ShipDef[Game.system.faction.policeShip]
+
+		-- create and equip them
+		while numberPolice > 0 do
+			local policeShip = Space.SpawnShipDocked(shiptype.id, self)
+			if policeShip == nil then
+				return
+			else
+				numberPolice = numberPolice - 1
+				--policeShip:SetLabel(Game.system.faction.policeName) -- this is cool, but not translatable right now
+				policeShip:SetLabel(l.POLICE)
+				policeShip:AddEquip(Equipment.laser.pulsecannon_dual_1mw)
+				policeShip:AddEquip(Equipment.misc.atmospheric_shielding)
+				policeShip:AddEquip(Equipment.misc.laser_cooling_booster)
+				policeShip:AddEquip(Equipment.cargo.hydrogen, 1)
+
+				table.insert(police[self], policeShip)
+			end
+		end
+	end
+
+	for _, policeShip in pairs(police[self]) do
+		-- if docked
+		if policeShip.flightState == "DOCKED" then
+			policeShip:Undock()
+		end
+		-- if not shot down
+		if policeShip:exists() then
+			policeShip:AIKill(targetShip)
+		end
+	end
+end
+
+
+--
+-- Method: LandPolice
+--
+-- Clear any target assigned and land flying station police.
+--
+-- > station:LandPolice()
+--
+--
+-- Availability:
+--
+--   2015 September
+--
+-- Status:
+--
+--   experimental
+--
+function SpaceStation:LandPolice()
+	-- land command issued before creation of police
+	if not police[self] then return end
+
+	for _, policeShip in pairs(police[self]) do
+		if not (policeShip.flightState == "DOCKED") and policeShip:exists() then
+			policeShip:CancelAI()
+			policeShip:AIDockWith(self)
+		end
+	end
 end
 
 
@@ -246,7 +407,9 @@ end
 -- Group: Methods
 --
 
-
+SpaceStation.lockedAdvert = nil
+SpaceStation.advertLockCount = 0
+SpaceStation.removeOnReleased = false
 SpaceStation.adverts = {}
 
 --
@@ -363,9 +526,12 @@ end
 --
 --  stable
 --
-
 function SpaceStation:RemoveAdvert (ref)
 	if not SpaceStation.adverts[self] then return end
+	if SpaceStation.lockedAdvert == ref then
+		SpaceStation.removeOnReleased = true
+		return
+	end
 	local onDelete = SpaceStation.adverts[self][ref].onDelete
 	if onDelete then
 		onDelete(ref)
@@ -374,6 +540,42 @@ function SpaceStation:RemoveAdvert (ref)
 	Event.Queue("onAdvertRemoved", self, ref)
 end
 
+--
+-- Method: LockAdvert
+--
+-- > station:LockAdvert(ref)
+--
+-- Only one advert may be locked at a time. Locked adverts are not removed by
+-- RemoveAdvert until they are unlocked.
+--
+-- Parameters:
+--
+--   ref - the advert reference number returned by <AddAdvert>
+--
+-- Availability:
+--
+--  September 2014
+--
+-- Status:
+--
+--  experimental
+--
+function SpaceStation:LockAdvert (ref)
+	if (SpaceStation.advertLockCount > 0) then
+		assert(SpaceStation.lockedAdvert == ref, "Attempt to lock ref "..ref
+		.."disallowed."
+		.." Ref "..(SpaceStation.lockedAdvert or "nil").." is already locked "
+		..SpaceStation.advertLockCount.." times.")
+		SpaceStation.advertLockCount = SpaceStation.advertLockCount + 1
+		return
+	end
+	assert(SpaceStation.lockedAdvert == nil, "Attempt to lock ref "..ref
+		.." disallowed."
+		.." Ref "..(SpaceStation.lockedAdvert or "nil").." is already locked.")
+	SpaceStation.lockedAdvert = ref
+	SpaceStation.removeOnReleased = false
+	SpaceStation.advertLockCount = 1
+end
 
 local function updateAdverts (station)
 	-- XXX this should really just be a single event
@@ -386,6 +588,43 @@ local function updateAdverts (station)
 	end
 end
 
+--
+-- Method: UnlockAdvert
+--
+-- > station:UnlockAdvert(ref)
+--
+-- Releases the preserved advert. There must be an advert preserved at the
+-- time. If RemoveAdvert(ref) was called with the preserved advert's ref while
+-- the advert was preserved, it is now removed.
+--
+-- Parameters:
+--
+--   ref - the advert reference number returned by <AddAdvert>
+--
+-- Availability:
+--
+--  September 2014
+--
+-- Status:
+--
+--  experimental
+--
+function SpaceStation:UnlockAdvert (ref)
+	assert(SpaceStation.lockedAdvert == ref, "Attempt to unlock ref "..ref
+		.." disallowed."
+		.." Ref "..(SpaceStation.lockedAdvert or "nil").." is locked"
+		..SpaceStation.advertLockCount.." times. Unlock this"
+		.." first.")
+	if (SpaceStation.advertLockCount > 1) then
+		SpaceStation.advertLockCount = SpaceStation.advertLockCount - 1
+		return
+	end
+	SpaceStation.lockedAdvert = nil
+	SpaceStation.advertLockCount = 0
+	if SpaceStation.removeOnReleased then
+		self:RemoveAdvert(ref)
+	end
+end
 
 local function updateSystem ()
 	local stations = Space.GetBodies(function (b) return b.superType == "STARPORT" end)
@@ -395,9 +634,19 @@ local function updateSystem ()
 		updateAdverts(stations[i])
 	end
 end
+
+local function createSystem()
+	local stations = Space.GetBodies(function (b) return b.superType == "STARPORT" end)
+	for i=1,#stations do
+		createShipMarket(stations[i])
+	end
+end
+
 local function destroySystem ()
 	equipmentStock = {}
 	equipmentPrice = {}
+
+	police = {}
 
 	shipsOnSale = {}
 
@@ -416,6 +665,7 @@ Event.Register("onGameStart", function ()
 	if (loaded_data) then
 		equipmentStock = loaded_data.equipmentStock
 		equipmentPrice = loaded_data.equipmentPrice or {} -- handle missing in old saves
+		police = loaded_data.police
 		for station,list in pairs(loaded_data.shipsOnSale) do
 			shipsOnSale[station] = {}
 			for i,entry in pairs(loaded_data.shipsOnSale[station]) do
@@ -433,11 +683,13 @@ Event.Register("onGameStart", function ()
 		loaded_data = nil
 	end
 
+	createSystem()
 	updateSystem()
 	Timer:CallEvery(3600, updateSystem)
 end)
 Event.Register("onEnterSystem", function (ship)
 	if ship ~= Game.player then return end
+	createSystem()
 	updateSystem()
 end)
 
@@ -445,6 +697,17 @@ Event.Register("onLeaveSystem", function (ship)
 	if ship ~= Game.player then return end
 	destroySystem()
 end)
+
+Event.Register("onShipDestroyed", function (ship, _)
+	for _,local_police in pairs(police) do
+		for k,police_ship in pairs(local_police) do
+			if (ship == police_ship) then
+				table.remove(local_police, k)
+			end
+		end
+	end
+end)
+
 Event.Register("onGameEnd", function ()
 	destroySystem()
 
@@ -452,6 +715,7 @@ Event.Register("onGameEnd", function ()
 	nextRef = 0
 	equipmentStock = {}
 	equipmentPrice = {}
+	police = {}
 	shipsOnSale = {}
 end)
 
@@ -461,6 +725,7 @@ Serializer:Register("SpaceStation",
 		local data = {
 			equipmentStock = equipmentStock,
 			equipmentPrice = equipmentPrice,
+			police = police,  --todo fails if a police ship is killed
 			shipsOnSale = {},
 		}
 		for station,list in pairs(shipsOnSale) do
